# docs/swagger.yaml
components:
  schemas:
...
    graph.NodeWithResources:
      properties:
        cores:
          type: integer
        cpuUsagePercent:
          type: number
        name:
          type: string
        ramTotalMB:
          type: number
        ramUsedMB:
          type: number
        ramUsagePercent:
          type: number
      type: object
...
    graph.CentralityServiceInfo:
      properties:
        centralityScore:
          type: number
        explanation:
          type: string
        name:
          type: string
        namespace:
          type: string
        riskLevel:
          type: string
        serviceId:
          type: string
      type: object
    graph.DataFreshness:
      properties:
        lastUpdatedSecondsAgo:
          type: integer
        source:
          type: string
        stale:
          type: boolean
        windowMinutes:
          type: integer
      type: object
    graph.TopCentralityResponse:
      properties:
        confidence:
          type: string
        dataFreshness:
          $ref: '#/components/schemas/graph.DataFreshness'
        metric:
          type: string
        services:
          items:
            $ref: '#/components/schemas/graph.CentralityServiceInfo'
          type: array
          uniqueItems: false
      type: object
    simulation.AddRiskAnalysis:
      properties:
        dependencyRisk:
          type: string
        description:
          type: string
      type: object
    simulation.AddSimulationRequest:
      properties:
        cpuRequest:
          type: number
        dependencies:
          items:
            $ref: '#/components/schemas/simulation.DependencyRef'
          type: array
          uniqueItems: false
        ramRequest:
          type: integer
        replicas:
          type: integer
        serviceName:
          type: string
        timeWindow:
          type: string
      type: object
    simulation.AddSimulationResult:
      properties:
        confidence:
          type: string
        explanation:
          type: string
        recommendation:
          $ref: '#/components/schemas/simulation.LegacyRecommendation'
        recommendations:
          items:
            $ref: '#/components/schemas/simulation.FailureRecommendation'
          type: array
          uniqueItems: false
        riskAnalysis:
          $ref: '#/components/schemas/simulation.AddRiskAnalysis'
        success:
          type: boolean
        suitableNodes:
          items:
            $ref: '#/components/schemas/simulation.NodeCapacity'
          type: array
          uniqueItems: false
        targetServiceName:
          type: string
        totalCapacityPods:
          type: integer
      type: object
    simulation.AffectedCaller:
      properties:
        edgeErrorRate:
          type: number
        lostTrafficRps:
          type: number
        name:
          type: string
        namespace:
          type: string
        serviceId:
          type: string
      type: object
...
    simulation.AffectedDownstream:
      properties:
        edgeErrorRate:
          type: number
        lostTrafficRps:
          type: number
        name:
          type: string
        namespace:
          type: string
        serviceId:
          type: string
      type: object
...
    simulation.BrokenPath:
      properties:
        path:
          items:
            type: string
          type: array
          uniqueItems: false
        pathRps:
          type: number
      type: object
    simulation.DataFreshness:
      properties:
        lastUpdatedSecondsAgo:
          type: integer
        source:
          type: string
        stale:
          type: boolean
        windowMinutes:
          type: integer
      type: object
    simulation.DependencyRef:
      properties:
        serviceId:
          type: string
      type: object
    simulation.FailureRecommendation:
      properties:
        action:
          type: string
        description:
          type: string
        priority:
          type: string
        reason:
          type: string
        target:
          type: string
        type:
          type: string
      type: object
    simulation.FailureSimulationRequest:
      properties:
        depth:
          type: integer
        serviceId:
          type: string
      type: object
    simulation.FailureSimulationResult:
      properties:
        affectedCallers:
          items:
            $ref: '#/components/schemas/simulation.AffectedCaller'
          type: array
          uniqueItems: false
        affectedDownstream:
          items:
            $ref: '#/components/schemas/simulation.AffectedDownstream'
          type: array
          uniqueItems: false
        confidence:
          type: string
        criticalPathsToTarget:
          items:
            $ref: '#/components/schemas/simulation.BrokenPath'
          type: array
          uniqueItems: false
        dataFreshness:
          $ref: '#/components/schemas/simulation.DataFreshness'
        explanation:
          type: string
        neighborhood:
          $ref: '#/components/schemas/simulation.NeighborhoodMeta'
        recommendations:
          items:
            $ref: '#/components/schemas/simulation.FailureRecommendation'
          type: array
          uniqueItems: false
        target:
          $ref: '#/components/schemas/simulation.ServiceRef'
        totalLostTrafficRps:
          type: number
        unreachableServices:
          items:
            $ref: '#/components/schemas/simulation.UnreachableService'
          type: array
          uniqueItems: false
      type: object
    simulation.LegacyRecommendation:
      properties:
        cpuRequest:
          type: number
        distribution:
          items:
            $ref: '#/components/schemas/simulation.PlacementDistribution'
          type: array
          uniqueItems: false
        ramRequest:
          type: integer
        serviceName:
          type: string
      type: object
    simulation.NeighborhoodMeta:
      properties:
        depthUsed:
          type: integer
        description:
          type: string
        edgeCount:
          type: integer
        generatedAt:
          type: string
        serviceCount:
          type: integer
      type: object
    simulation.NodeCapacity:
      properties:
        availableCpu:
          type: number
        availableRam:
          type: number
        canFit:
          type: boolean
        cpuAvailable:
          type: number
        cpuTotal:
          type: number
        maxPods:
          type: integer
        node:
          type: string
        nodeName:
          type: string
        ramAvailableMB:
          type: number
        ramTotalMB:
          type: number
        reason:
          type: string
        score:
          type: integer
        suitable:
          type: boolean
      type: object
...
    simulation.ScalingModel:
      properties:
        alpha:
          type: number
        type:
          type: string
      type: object
    simulation.ScalingSimulationRequest:
      properties:
        currentPods:
          type: integer
        latencyMetric:
          type: string
        maxDepth:
          type: integer
        model:
          $ref: '#/components/schemas/simulation.ScalingModel'
        newPods:
          type: integer
        serviceId:
          type: string
        timeWindow:
          type: string
      type: object
    simulation.ScalingSimulationResult:
      properties:
        affectedCallers:
          $ref: '#/components/schemas/simulation.AffectedCallersList'
        affectedPaths:
          items:
            $ref: '#/components/schemas/simulation.AffectedPathScaling'
          type: array
          uniqueItems: false
        confidence:
          type: string
        currentPods:
          type: integer
        dataFreshness:
          $ref: '#/components/schemas/simulation.DataFreshness'
        explanation:
          type: string
        latencyEstimate:
          $ref: '#/components/schemas/simulation.ScalingLatencyEstimate'
        latencyMetric:
          type: string
        neighborhood:
          $ref: '#/components/schemas/simulation.NeighborhoodMeta'
        newPods:
          type: integer
        recommendations:
          items:
            $ref: '#/components/schemas/simulation.FailureRecommendation'
          type: array
          uniqueItems: false
        scalingDirection:
          type: string
        scalingModel:
          $ref: '#/components/schemas/simulation.ScalingModel'
        target:
          $ref: '#/components/schemas/simulation.ServiceRef'
        warnings:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
    simulation.ServiceRef:
      properties:
        name:
          type: string
        namespace:
          type: string
        serviceId:
          type: string
      type: object
    simulation.UnreachableService:
      properties:
        lostFromReachableCutsRps:
          type: number
        lostFromTargetRps:
          type: number
        lostTrafficRps:
          type: number
        name:
          type: string
        namespace:
          type: string
        serviceId:
          type: string
      type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for the Predictive Analysis Engine, responsible for failure simulation,
    scaling analysis, and risk assessment.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Predictive Analysis Engine API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /infrastructure/nodes:
    get:
      description: Fetches current infrastructure nodes and their resource usage
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/graph.NodeWithResources'
                type: array
                uniqueItems: false
          description: OK
        "503":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Service Unavailable
      summary: Get Infrastructure Nodes
      tags:
      - infrastructure
...
  /risk/services/top:
    get:
      description: Returns services ordered by risk metrics (pagerank or betweenness)
      parameters:
      - description: Risk metric (pagerank, betweenness)
        in: query
        name: metric
        schema:
          default: pagerank
          type: string
      - description: Number of services to return (1-20)
        in: query
        name: limit
        schema:
          default: 5
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/graph.TopCentralityResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Service Unavailable
      summary: Get Top Risky Services
      tags:
      - risk
  /services:
    get:
      description: Fetches a list of services and their status from the Graph Engine
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: OK
        "503":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Service Unavailable
      summary: List Services
      tags:
      - services
  /simulate/add:
    post:
      description: Simulates adding a new service to the cluster (capacity planning)
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/simulation.AddSimulationRequest'
                description: Simulation parameters
                summary: request
        description: Simulation parameters
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/simulation.AddSimulationResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Simulate Adding Service
      tags:
      - simulation
  /simulate/failure:
    post:
      description: Simulates a failure of a specific service and analyzes the impact
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/simulation.FailureSimulationRequest'
                description: Simulation parameters
                summary: request
        description: Simulation parameters
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/simulation.FailureSimulationResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Service Unavailable
      summary: Simulate Service Failure
      tags:
      - simulation
  /simulate/scale:
    post:
      description: Simulates scaling a service and analyzes latency impact
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/simulation.ScalingSimulationRequest'
                description: Simulation parameters
                summary: request
        description: Simulation parameters
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/simulation.ScalingSimulationResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Service Unavailable
      summary: Simulate Scaling
      tags:
      - simulation
...